requirements.txt
Django>=5.0,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
django-ratelimit>=4.1,<5.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
# Uncomment if using PostgreSQL instead of SQLite
# psycopg2-binary>=2.9,<3.0
